requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
Pillow>=10.0.0
requests>=2.31.0
asyncpg>=0.29.0